Remove call to secureArea.getPublicKey during openID4vp presentation of SD-JWT document. Refactor SD-JWT validation and parsing in SdJwtUtils and OpenId4VpService.
