+++ Sources/EudiWalletKit/Services/OpenId4VciService.swift
@@ -1040,15 +1040,7 @@ public actor OpenId4VciService {
 	}
 
 	private func validateSdJwtBindingKeys(_ serialized: String, publicCoseKeys: inout [CoseKey]) throws {
-		let (_, payload, _) = SdJwtUtils.extractJWTParts(serialized)
-		guard let payloadData = Data(base64URLEncoded: payload) else {
-			throw PresentationSession.makeError(str: "Failed to decode SD-JWT payload")
-		}
-		let payloadJson = try JSON(data: payloadData)
-		guard payloadJson["cnf"].exists(), payloadJson["cnf"].type == .dictionary else {
-			throw PresentationSession.makeError(str: "Issued SD-JWT is missing a valid cnf claim")
-		}
-		let cnfKeys = try parseCnfBindingKeys(payloadJson["cnf"])
+		let cnfKeys = try SdJwtUtils.parseCnfBindingKeys(fromSerializedCredential: serialized)
 		let availableKeys = publicCoseKeys.map(\.x963Representation)
 		for key in cnfKeys {
 			guard let x = Data(base64URLEncoded: key.x), let y = Data(base64URLEncoded: key.y) else {
@@ -1061,31 +1053,6 @@ public actor OpenId4VciService {
 			} else {
 				throw PresentationSession.makeError(str: "Failed to find matching public key for SD-JWT cnf binding key")
 			}
-		}
-	}
-
-	private func parseCnfBindingKeys(_ cnf: JSON) throws -> [ECPublicKey] {
-		var jwks: [JSON] = []
-		if cnf["jwk"].type == .dictionary {
-			jwks.append(cnf["jwk"])
-		} else if cnf["jwk"].type == .array {
-			jwks.append(contentsOf: cnf["jwk"].arrayValue)
-		}
-		if cnf["jwks"]["keys"].type == .array {
-			jwks.append(contentsOf: cnf["jwks"]["keys"].arrayValue)
-		}
-		guard !jwks.isEmpty else {
-			throw PresentationSession.makeError(str: "Issued SD-JWT cnf claim does not contain JWK binding keys")
-		}
-		return try jwks.map { jwk in
-			guard let kty = jwk["kty"].string, kty == "EC",
-					let curveName = jwk["crv"].string,
-					let curve = ECCurveType(rawValue: curveName),
-					let x = jwk["x"].string,
-					let y = jwk["y"].string else {
-				throw PresentationSession.makeError(str: "Issued SD-JWT cnf JWK is missing required key material")
-			}
-			return ECPublicKey(crv: curve, x: x, y: y)
 		}
 	}
